% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/control.utilities.R
\name{as.control.list}
\alias{as.control.list}
\alias{as.control.list.control.list}
\alias{as.control.list.list}
\title{Convert to a control list.}
\usage{
as.control.list(x, ...)

\method{as.control.list}{control.list}(x, ...)

\method{as.control.list}{list}(x, FUN = NULL, ...)
}
\arguments{
\item{x}{An object, usually a \code{\link{list}}, to be converted to a
control list.}

\item{FUN}{Either a \verb{control.*()} function or its name or suffix
(to which \code{"control."} will be prepended); defaults to taking the
nearest (in the call traceback) function that does not begin with
\code{"as.control.list"}, and prepending \code{"control."} to it. (This is
typically the function that called \code{as.control.list()} in the
first place.)}
}
\value{
a \code{control.list} object.
}
\description{
Convert to a control list.
}
\section{Methods (by class)}{
\itemize{
\item \code{control.list}: Idempotent method for control lists.

\item \code{list}: The method for plain lists, which runs
them through \code{FUN}.
}}

\examples{
myfun <- function(..., control=control.myfun()){
  as.control.list(control)
}
control.myfun <- function(a=1, b=a+1){
  list(a=a,b=b)
}

myfun()
myfun(control = list(a=2))
}
