% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rle_utils.R
\name{as.rle}
\alias{as.rle}
\alias{as.rle.rle}
\alias{as.rle.default}
\title{Coerce to \code{\link{rle}} if not already an \code{\link{rle}} object.}
\usage{
as.rle(x)

\method{as.rle}{rle}(x)

\method{as.rle}{default}(x)
}
\arguments{
\item{x}{the object to be coerced.}
}
\description{
Coerce to \code{\link{rle}} if not already an \code{\link{rle}} object.
}
