% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/misc.utilities.R
\name{attr}
\alias{attr}
\title{A wrapper for base::attr which defaults to exact matching.}
\usage{
attr(x, which, exact = TRUE)
}
\arguments{
\item{x, which, exact}{as in \code{base::attr}, but with \code{exact}
defaulting to \code{TRUE} in this implementation}
}
\value{
as in \code{base::attr}
}
\description{
A wrapper for base::attr which defaults to exact matching.
}
\examples{

x <- list()
attr(x, "name") <- 10

base::attr(x, "n")

stopifnot(is.null(attr(x, "n")))

base::attr(x, "n", exact = TRUE)
}
