% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/matrix.utils.R
\name{cols_constant}
\alias{cols_constant}
\title{Identify Matrix Columns with a Small Range}
\usage{
cols_constant(x, tol = .Machine$double.eps * 2)
}
\arguments{
\item{x}{a \code{\link{matrix}} or a list of matrices, such as an
\code{\link[coda:mcmc.list]{mcmc.list}}.}

\item{tol}{a vector (recycled as needed) giving maximum half-ranges
allowed.}
}
\value{
A logical vector.
}
\description{
Identify Matrix Columns with a Small Range
}
\examples{
M <- cbind(rnorm(5), rnorm(10, 1, 1e-16))
M
sweep(M, 2L, M[1L, ])
(sweep(M, 2L, M[1L, ]) == 0) |> apply(2, all)
cols_constant(M)

}
