% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/misc.utilities.R
\name{deInf}
\alias{deInf}
\title{Truncate values of high magnitude in a vector.}
\usage{
deInf(x, replace = 1/.Machine$double.eps)
}
\arguments{
\item{x}{a numeric or integer vector.}

\item{replace}{a number or a string \code{"maxint"} or \code{"intmax"}.}
}
\value{
Returns \code{x} with elements whose magnitudes exceed \code{replace}
replaced replaced by \code{replace} (or its negation). If \code{replace} is
\code{"maxint"} or \code{"intmax"}, \code{.Machine$integer.max} is used instead.

\code{NA} and \code{NAN} values are preserved.
}
\description{
Truncate values of high magnitude in a vector.
}
