% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/string.utilities.R
\name{despace}
\alias{despace}
\title{A one-line function to strip whitespace from its argument.}
\usage{
despace(s)
}
\arguments{
\item{s}{a character vector.}
}
\description{
A one-line function to strip whitespace from its argument.
}
\examples{
stopifnot(despace("\n \t  ")=="")
}
