% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/formula.utilities.R
\name{envir}
\alias{envir}
\alias{envir<-}
\title{A generic for querying and setting an object's environment}
\usage{
envir(object)

envir(object) <- value
}
\arguments{
\item{object}{object whose environment is to be queried or set}

\item{value}{typically an \code{\link{environment}}, but could be any RHS
supported by the method}
}
\description{
\code{\link[=environment]{environment()}} and \code{\link[=environment<-]{environment<-()}} are not generics, so it is
not possible to dispatch based on the class of the object affected.
}
\details{
When no method is available, these generics fall back to the
\code{\link[=environment]{environment()}} and \code{\link[=environment<-]{environment<-()}} functions.
}
