% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/misc.utilities.R
\name{hasAttr}
\alias{hasAttr}
\title{Check if an object has an attribute with a specified name.}
\usage{
hasAttr(x, name)
}
\arguments{
\item{x}{Any object.}

\item{name}{One or more character values to look for.}
}
\value{
A logical vector of the same length as \code{name}, containing
\code{TRUE} if \code{attr(x, name, exact = TRUE)} is not \code{NULL}.
}
\description{
Analogous function to \code{\link[utils:hasName]{utils::hasName()}} but for \code{\link[=attr]{attr()}}
attributes.
}
\examples{
x <- structure(list(), name = 10)

stopifnot(hasAttr(x, c("n", "name")) == c(FALSE, TRUE))

}
\seealso{
\code{\link[=hasName]{hasName()}}
}
