% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/matrix.utils.R
\name{is.SPD}
\alias{is.SPD}
\title{Test if the object is a matrix that is symmetric and positive definite}
\usage{
is.SPD(x, tol = .Machine$double.eps)
}
\arguments{
\item{x}{the object to be tested.}

\item{tol}{the tolerance for the reciprocal condition number.}
}
\description{
Test if the object is a matrix that is symmetric and positive definite
}
