% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/matrix.utils.R
\name{matrix_helpers}
\alias{matrix_helpers}
\alias{sweep_cols.matrix}
\alias{sweep.matrix}
\alias{apply.matrix}
\title{Specialization of Common Functions to Matrices}
\usage{
sweep_cols.matrix(x, STATS, disable_checks = FALSE)

sweep.matrix(x, MARGIN, STATS, FUN = "-", check.margin = TRUE, ...)

apply.matrix(x, MARGIN, FUN, ..., simplify = TRUE)
}
\arguments{
\item{x}{a matrix.}

\item{disable_checks}{if \code{TRUE}, do not check that \code{x} is a
numeric matrix and its number of columns matches the length of \code{STATS};
set in production code for a significant speed-up.}

\item{MARGIN, FUN, simplify, check.margin, STATS, ...}{See help for corresponding function.}
}
\description{
Being able to assume two dimensions reduces overhead.
}
\section{Functions}{
\itemize{
\item \code{sweep_cols.matrix()}: Highly optimized \code{\link{sweep}}\verb{(x, 2, STATS, "-")} specifically for numeric matrices.

\item \code{sweep.matrix()}: \code{\link{sweep}}\verb{(x, 2L, STATS, FUN, ...)},
though for convenience, STATS can also be a function that is
applied to each column; this also disables the margin check by default.

\item \code{apply.matrix()}: \code{\link{apply}}\verb{(x, 2L, FUN, ...)}.

}}
\examples{

x <- matrix(runif(1000), ncol = 4)

s <- 1:4

stopifnot(all.equal(sweep_cols.matrix(x, s), sweep(x, 2, s)))

stopifnot(all.equal(sweep(x, 2, colMeans(x)), sweep.matrix(x, 2, colMeans(x))))
stopifnot(all.equal(sweep(x, 2, colMeans(x)), sweep.matrix(x, 2, mean)))
stopifnot(all.equal(apply.matrix(x, 2, min), apply(x, 2, min)))
}
