% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/misc.utilities.R
\name{NVL}
\alias{NVL}
\title{Return the first argument passed (out of any number) that is not
\code{NULL}.}
\usage{
NVL(...)
}
\arguments{
\item{\dots}{Expressions to be tested.}
}
\value{
The first argument that is not \code{NULL}, or \code{NULL} if all
are.
}
\description{
This function is inspired by SQL function \code{NVL}, and simply returns the
first argument that is not \code{NULL}, or \code{NULL} if all arguments are
NULL.
}
\details{
Note that an earlier version of this function took only two arguments:
\code{EXPR}, that was tested and returned if not \code{NULL} and
\code{NULLV}, which was returned if \code{EXPR} was \code{NULL}. The new
version produces identical output for the same (two-argument) input, but
tests any number of expressions sequentially.
}
\examples{

a <- NULL

print(a) # NULL
print(NVL(a,0)) # 0

b <- 1

print(b) # 1
print(NVL(b,0)) # 1

# Also,
print(NVL(NULL,1,0)) # 1
print(NVL(NULL,0,1)) # 0
print(NVL(NULL,NULL,0)) # 0
print(NVL(NULL,NULL,NULL)) # NULL
}
\seealso{
\code{\link[base]{is.null}}, \code{\link[base]{if}}
}
\keyword{utilities}
