% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/matrix.utils.R
\name{pdet_rat}
\alias{pdet_rat}
\title{(Pseudo-)Determinant of the ratio of two matrices}
\usage{
pdet_rat(num, denom, log = FALSE, root = FALSE, ...)
}
\arguments{
\item{num, denom}{numerator and denominator matrices.}

\item{log}{whether to return log-pseudo-determinant.}

\item{root}{whether to return \eqn{p}th root of the ratio, where
\eqn{p} is the effective rank.}

\item{...}{additional arguments to \code{\link[=qrssolve]{qrssolve()}}, particularly \code{tol}
for determining the cut-off for estimating the rank and \code{snnd}
for whether \code{denom} can be assumed to be symmetric and positive
non-negative definite.}
}
\value{
The pseudo-determinant, with an additional attribute
\code{"rank"} giving the number of eigenvalues used.
}
\description{
(Pseudo-)Determinant of the ratio of two matrices
}
\note{
Kernel of \code{denom} must be contained in the kernel of \code{num},
or equivalently, the span of \code{num} must be contained in the span
of \code{denom}.
}
