%  File man/print.control.list.Rd in package statnet.common, part of the Statnet suite
%  of packages for network analysis, http://statnet.org .
%
%  This software is distributed under the GPL-3 license.  It is free,
%  open source, and has the attribution requirements (GPL Section 7) at
%  http://statnet.org/attribution
%
%  Copyright 2003-2013 Statnet Commons
%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%
\name{print.control.list}
\alias{print.control.list}
\title{
  Pretty print the control list
}
\description{
  This function prints the control list, including what it can control
  and the elements. 
}
\usage{
\method{print}{control.list}(x, ...)
}
\arguments{
  \item{x}{
    A list generated by a \code{control.*} function.
}
  \item{\dots}{
    Unused at this time.
}
}
\seealso{
  check.control.class, set.control.class
}

\keyword{ utilities }
