% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/control.utilities.R
\name{sctrl}
\alias{sctrl}
\title{Statnet Control: a utility to facilitate argument completion of control lists.}
\usage{
sctrl(...)
}
\arguments{
\item{...}{Arguments to be put into the list. They \emph{must} be named.}
}
\description{
In and of itself, \code{sctrl} copies its named arguments into a
list. However, its argument list is updated dynamically as packages
are loaded, as are those of its reexports from other packages. This
is done using an API provided by helper functions. (See \code{API?sctrl}.)
}
\note{
You may see messages along the lines of\if{html}{\out{<div class="The following object is masked from 'package:PKG':">}}\preformatted{sctrl
}\if{html}{\out{</div>}}

when loading packages. They are benign.
}
