% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/formula.utilities.R
\name{sign<-}
\alias{sign<-}
\title{A generic for setting the sign of an object}
\usage{
sign(x) <- value
}
\arguments{
\item{x}{object whose sign is to be set}

\item{value}{a numeric vector specifying the sign}
}
\description{
A generic for setting the sign of an object
}
