% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/test.utilities.R
\name{skip_if_not_checking}
\alias{skip_if_not_checking}
\title{Skip a test if not called via \verb{R CMD} \code{\link{check}}}
\usage{
skip_if_not_checking()
}
\description{
A \CRANpkg{testthat} predicate to skip tests if not run as a part
of a package \link{check}.
}
