% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/control.utilities.R
\name{snctrl}
\alias{snctrl}
\title{Statnet Control}
\usage{
snctrl(...)
}
\arguments{
\item{...}{The parameter list is updated dynamically as packages
are loaded and unloaded. Their current list is given below.}
}
\description{
A utility to facilitate argument completion of control lists.
}
\details{
In and of itself, \code{snctrl} copies its named arguments into a
list. However, its argument list is updated dynamically as packages
are loaded, as are those of its reexports from other packages. This
is done using an API provided by helper functions. (See \code{API?snctrl}.)
}
\note{
You may see messages along the lines of

\if{html}{\out{<div class="sourceCode">}}\preformatted{The following object is masked from 'package:PKG':
snctrl
}\if{html}{\out{</div>}}

when loading packages. They are benign.
}
\section{Currently recognised control parameters}{

This list is updated as packages are loaded and unloaded.

\Sexpr[results=rd,stage=render]{statnet.common::snctrl_names()}
}

