% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/statnet.common-deprecated.R
\name{statnet.common-deprecated}
\alias{statnet.common-deprecated}
\alias{compact.rle}
\title{Deprecated functions from \code{statnet.common}}
\usage{
compact.rle(...)
}
\arguments{
\item{...}{arguments to deprecated functions.}
}
\description{
Deprecated functions from \code{statnet.common}
}
