% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/misc.utilities.R
\name{vector.namesmatch}
\alias{vector.namesmatch}
\title{reorder vector v into order determined by matching the names of its elements
to a vector of names}
\usage{
vector.namesmatch(v, names, errname = NULL)
}
\arguments{
\item{v}{a vector (or list) with named elements, to be reorderd}

\item{names}{a character vector of element names, corresponding to names of
\code{v}, specificying desired orering of \code{v}}

\item{errname}{optional, name to be reported in any error messages. default
to \code{deparse(substitute(v))}}
}
\value{
returns \code{v}, with elements reordered
}
\description{
This function is deprecated in favor of \code{\link[=match_names]{match_names()}} and will be
removed in a future release.
}
\note{
earlier versions of this function did not order as advertised
}
\examples{

test<-list(c=1,b=2,a=3)
vector.namesmatch(test,names=c('a','c','b'))
}
