% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/misc.utilities.R
\name{all_identical}
\alias{all_identical}
\title{Test if all items in a vector or a list are identical.}
\usage{
all_identical(x)
}
\arguments{
\item{x}{a vector or a list}
}
\value{
\code{TRUE} if all elements of \code{x} are identical to each other.
}
\description{
Test if all items in a vector or a list are identical.
}
\examples{

stopifnot(!all_identical(1:3))

stopifnot(all_identical(list("a", "a", "a")))
}
\seealso{
\code{\link[base:identical]{base::identical()}}
}
