% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/control.utilities.R
\name{print.control.list}
\alias{print.control.list}
\title{Pretty print the control list}
\usage{
\method{print}{control.list}(x, ...)
}
\arguments{
\item{x}{A list generated by a \code{control.*} function.}

\item{\dots}{Unused at this time.}
}
\description{
This function prints the control list, including what it can control and the
elements.
}
\seealso{
\code{\link{check.control.class}}, \code{\link{set.control.class}}
}
\keyword{utilities}
