\name{print.control.list}
\alias{print.control.list}
\title{
  Pretty print the control list
}
\description{
  This function prints the control list, including what it can control
  and the elements. 
}
\usage{
\method{print}{control.list}(x, ...)
}
\arguments{
  \item{x}{
    A list generated by a \code{control.*} function.
}
  \item{\dots}{
    Unused at this time.
}
}
\seealso{
  check.control.class, set.control.class
}

\keyword{ utilities }
