% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/control.utilities.R
\name{sctrl}
\alias{sctrl}
\title{Statnet Control: a utility to facilitate argument completion of control lists.}
\usage{
sctrl(...)
}
\arguments{
\item{...}{The parameter list is updated dynamically as packages
are loaded and unloaded. Their current list is given below.}
}
\description{
In and of itself, \code{sctrl} copies its named arguments into a
list. However, its argument list is updated dynamically as packages
are loaded, as are those of its reexports from other packages. This
is done using an API provided by helper functions. (See \code{API?sctrl}.)
}
\note{
You may see messages along the lines of\preformatted{The following object is masked from 'package:PKG':
sctrl
}

when loading packages. They are benign.
}
\section{Currently loaded}{

This is a dynamically updated list.

\Sexpr[results=rd,stage=render]{statnet.common:::.sctrl_names()}
}

