% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/control.utilities.R
\name{set.control.class}
\alias{set.control.class}
\title{Set the class of the control list}
\usage{
set.control.class(myname = as.character(ult(sys.calls(), 2)[[1]]),
  control = get("control", pos = parent.frame()))
}
\arguments{
\item{myname}{Name of the class to set.}

\item{control}{Control list. Defaults to the \code{control} variable in the
calling function.}
}
\value{
The control list with class set.
}
\description{
This function sets the class of the control list, with the default being the
name of the calling function.
}
\note{
In earlier versions, \code{OKnames} and \code{myname} were autodetected. This capability has been deprecated and results in a warning issued once per session. They now need to be set explicitly.
}
\seealso{
check.control.class, print.control.list
}
\keyword{utilities}
